+++ src/pato/models/response.py
@@ -197,8 +197,8 @@ class DataCollectionSummary(BaseDataCollectionOut):
     nominalDefocus: Optional[float] = None
     phasePlate: Optional[str] = None
     dataCollectionPlanId: Optional[int] = None
-    tomograms: int
     globalAlignmentQuality: Optional[float] = None
+    tomograms: int
 
     @field_validator("phasePlate", mode="before")
     def to_bool_str(cls, v):
@@ -496,3 +496,7 @@ class FoilHole(BaseModel):
     x: int = Field(validation_alias="pixelLocationX")
     y: int = Field(validation_alias="pixelLocationY")
     movieCount: Optional[int] = 0
+    particleCount: Optional[float] = None
+    resolution: Optional[float] = None
+    defocus: Optional[float] = None
+    astigmatism: Optional[float] = None