+++ tests/autoproc/test_resolution.py
@@ -3,7 +3,7 @@ def test_get_user(mock_permissions, client):
     resp = client.get("/autoProc/56986680/resolution?dataBin=5")
     assert resp.status_code == 200
     assert resp.json()["items"][2]["x"] == 5
-    assert resp.json()["items"][2]["y"] == 3
+    assert resp.json()["items"][2]["y"] == 4
 
 
 def test_get_minimum(mock_permissions, client):
@@ -12,7 +12,7 @@ def test_get_minimum(mock_permissions, client):
     resp = client.get("/autoProc/56986680/resolution?minimum=900")
     assert resp.status_code == 200
     assert resp.json()["items"][0]["x"] == "<900.0"
-    assert resp.json()["items"][0]["y"] == 5
+    assert resp.json()["items"][0]["y"] == 7
 
 
 def test_not_found(mock_permissions, client):